first virsion of working monitor to be tested

+++ cli/image_tools.py
@@ -6,8 +6,8 @@ from PIL import Image
 import cv2 as cv2
 
 
-def makeThumb(path, thumbpath):
-  return makeThumb_opencv(path, thumbpath)
+def makeThumb(path, thumbpath, overwrite):
+  return makeThumb_opencv(path, thumbpath, overwrite)
 
 def makeThumb_pillow(path, thumbdir):
 
@@ -32,22 +32,23 @@ def makeThumb_pillow(path, thumbdir):
   # save thumb
   im.save(thumbfile_with_ext)
 
-def makeThumb_opencv(path, thumbpath):
-
+def makeThumb_opencv(path, thumbpath, overwrite):
 
   # replace old ext with png
   thumbpath_with_ext = os.path.splitext(thumbpath)[0]+'.png'
 
-  # create thumb
-  maxsize = (120,120)
+  if overwrite or (not os.path.isfile(thumbpath_with_ext)):
 
-  origimg = cv2.imread(path)
-  imRes = cv2.resize(origimg, maxsize, interpolation = cv2.INTER_AREA)
+    # create thumb
+    maxsize = (120,120)
 
-  # create dir if needed
-  directory = os.path.dirname(thumbpath_with_ext)
-  if not os.path.exists(directory):
-    os.makedirs(directory)
+    origimg = cv2.imread(path)
+    imRes = cv2.resize(origimg, maxsize, interpolation = cv2.INTER_AREA)
 
-  # save thumb
-  cv2.imwrite(thumbpath_with_ext, imRes)
+    # create dir if needed
+    directory = os.path.dirname(thumbpath_with_ext)
+    if not os.path.exists(directory):
+      os.makedirs(directory)
+
+    # save thumb
+    cv2.imwrite(thumbpath_with_ext, imRes)
